pyproject.toml
[tool.poetry]
name = "ezpanso"
version = "1.2.1"
description = "A modern GUI for managing Espanso text expansion snippets"
authors = ["Longman <info@longman.hk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
PyQt6 = "^6.5.0"
PyYAML = "^6.0"
"ruamel.yaml" = "^0.18.0"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.13.0"
pytest = "^8.0.0"
pytest-qt = "^4.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ezpanso = "main:main"
